Rename projects to be consistent (#605)
While at it, Remove and Sort usings in touched files with the option of
`dotnet_sort_system_directives_first = false`.
