+++ src/backend/PortabilityService.Functions/PortabilityService.Functions.csproj
@@ -10,8 +10,8 @@
   <ItemGroup>
     <ProjectReference Include="..\..\lib\Microsoft.Fx.Portability.Azure\Microsoft.Fx.Portability.Azure.csproj" />
     <ProjectReference Include="..\..\lib\Microsoft.Fx.Portability\Microsoft.Fx.Portability.csproj" />
-    <ProjectReference Include="..\PortabilityService.DependencyInjection\PortabilityService.DependencyInjection.csproj" />
-    <ProjectReference Include="..\WorkflowManagement\WorkflowManagement.csproj" />
+    <ProjectReference Include="..\PortabilityService.Functions.DependencyInjection\PortabilityService.Functions.DependencyInjection.csproj" />
+    <ProjectReference Include="..\PortabilityService.WorkflowManagement\PortabilityService.WorkflowManagement.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="Microsoft.CSharp" />
