+++ src/backend/PortabilityService.Functions/ProcessWorkflowQueue.cs
@@ -1,13 +1,12 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
-using WorkflowManagement;
+using PortabilityService.WorkflowManagement;
+using System.Threading.Tasks;
 
-namespace Functions
+namespace PortabilityService.Functions
 {
     public static class ProcessWorkflowQueue
     {
