+++ src/backend/PortabilityService.Functions/local.settings.json
