+++ src/backend/PortabilityService.WorkflowManagement/WorkflowActions/TelemetryAction.cs
@@ -3,7 +3,7 @@
 
 using System.Threading.Tasks;
 
-namespace WorkflowManagement
+namespace PortabilityService.WorkflowManagement
 {
     class TelemetryAction : IWorkflowAction
     {
