+++ src/backend/PortabilityService.WorkflowManagement/WorkflowQueueMessage.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace WorkflowManagement
+namespace PortabilityService.WorkflowManagement
 {
     public class WorkflowQueueMessage
     {
